requirements.txt
streamlit
tensorflow>=2.20
keras
numpy
pillow
gdown

